fix :butterfly: Rename RoomGame.ts to RoomGameData.ts
