+++ functions/src/types/request/bodyData/index.ts
@@ -2,5 +2,5 @@ export * from "./LoginData";
 export * from "./UserData";
 export * from "./GameData";
 export * from "./RoomData";
-export * from "./RoomGame";
+export * from "./RoomGameData";
 export * from "./EventData";
